FEAT: Semantic Embedding on pages using pgVector

+++ SiteBackend/Data/SeedData/LoadSeedData.cs
@@ -1,5 +1,8 @@
 using System.Collections.Concurrent;
+using System.Diagnostics;
+using EFCore.BulkExtensions;
 using SiteBackend.Database;
+using SiteBackend.Models.SearchEngine;
 using SiteBackend.Models.SearchEngine.Index;
 
 namespace SiteBackend.Data.SeedData;
@@ -8,9 +11,11 @@ public static class LoadSeedData
 {
     // TODO: not this.
     private static readonly string SiteSeedPath = "/home/zachary/Downloads/seed.csv";
+    private static readonly string WordSeedPath = "/home/zachary/Downloads/words.txt";
     public static void SeedDatabase(SearchEngineCtx dbCtx, bool isDevelopment = true)
     {
         List<string[]> sites = [];
+        string[] words = new string[] { };
 
         if (!File.Exists(SiteSeedPath))
         {
@@ -27,6 +32,10 @@ public static class LoadSeedData
                 var values = line.Split(','); 
                 sites.Add(values);
             }
+            Console.WriteLine($"Loaded {sites.Count} sites from {SiteSeedPath}");
+            
+            words = File.ReadAllLines(WordSeedPath);
+            Console.WriteLine($"Loaded {words.Length} words");
         }
         catch (Exception ex)
         {
@@ -38,8 +47,9 @@ public static class LoadSeedData
         {
             Console.Write("WARNING - WARNING ");
             Console.WriteLine();
-            Console.WriteLine($"PRESS ANY KEY TO DELETE DATABASE CONTINUE TO THE DEVELOPMENT");
-            Console.ReadKey();
+            Console.WriteLine($"Press Y to wipe db");
+            var resp = Console.ReadKey();
+            if (resp.Key != ConsoleKey.Y) return;
             Console.WriteLine("Deleting Database...");
             dbCtx.Database.EnsureDeleted();
             Console.WriteLine("Database deleted...");
@@ -47,26 +57,61 @@ public static class LoadSeedData
             dbCtx.Database.EnsureCreated();
             Console.WriteLine("Database created...");
             Console.WriteLine("Getting Seed Sites...");
-            sites = sites.Take(sites.Count / 400).ToList(); // otherwise, will take longer than heat-death of universe.
+            
+            sites = sites.Take(sites.Count / 500).ToList();
         }
 
-        ConcurrentBag<Website> websites = [];
-        Parallel.ForEach(sites, site =>
+        var websites = new ConcurrentBag<Website>();
+        var dictionary = new ConcurrentBag<Word>();
+        Parallel.Invoke(
+            () =>
             {
-                string trimmedSite = site[1].Trim('"');
-                if (Uri.TryCreate(trimmedSite, UriKind.RelativeOrAbsolute, out Uri _))
-                    websites.Add(new Website(trimmedSite));
-                else
-                    Console.WriteLine($"Error: Invalid URL: {trimmedSite}");
+                Parallel.ForEach(sites, new ParallelOptions { MaxDegreeOfParallelism = Environment.ProcessorCount }, 
+                    site =>
+                    {
+                        string trimmedUrl = site[1].Trim('"');
+                        var newSite = new Website(trimmedUrl);
+                        if (Uri.TryCreate(trimmedUrl, UriKind.RelativeOrAbsolute, out Uri _))
+                            websites.Add(newSite);
+                        else
+                            Console.WriteLine($"Site {trimmedUrl} could not be parsed into a valid URL.");
+                        
+                        Debug.Assert(newSite.Sitemap != null, "sitemap null...");
+                    });
+            },
+            () =>
+            {
+                Parallel.ForEach(words, new ParallelOptions { MaxDegreeOfParallelism = Environment.ProcessorCount }, 
+                    word =>
+                    {
+                        dictionary.Add(new Word { Text = word });
+                    });
             }
         );
         
-        dbCtx.ChangeTracker.AutoDetectChangesEnabled = false;
-        Console.WriteLine($"Loaded {websites.Count} websites");
-        dbCtx.Websites.AddRange(websites);
-        Console.WriteLine($"Detecting changes for {websites.Count} websites...");
-        dbCtx.ChangeTracker.DetectChanges();
-        Console.WriteLine($"Saving {websites.Count} websites...");
+        
+        var bulkConfig = new BulkConfig
+        {
+            PreserveInsertOrder = false,
+            SetOutputIdentity = false,
+            BatchSize = 5000,
+            BulkCopyTimeout = 0,
+            EnableStreaming = true,
+            TrackingEntities = false,
+            WithHoldlock = false,
+            IncludeGraph = true,
+        };
+
+        dbCtx.Database.EnsureCreated();
+        Console.WriteLine($"Inserting {dictionary.Count} Words...");
+        bulkConfig.BatchSize = 50000;
+        dbCtx.BulkInsert(dictionary, bulkConfig);
+        dbCtx.SaveChanges();
+        Console.WriteLine($"Inserting {websites.Count} Websites...");
+        bulkConfig.BatchSize = 50000;
+        dbCtx.BulkInsert(websites, bulkConfig);
+        
+        Console.WriteLine($"Saving {websites.Count + dictionary.Count} Seed Items...");
         dbCtx.SaveChanges();
     }
 }
